chore (refs T22614): move core store to client/js.
