+++ demosplan/DemosPlanCoreBundle/Resources/client/js/InitVue.js
@@ -11,7 +11,7 @@ import { bootstrap } from '@DpJs/bootstrap'
 import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpNotifyContainer from '@DpJs/components/core/notify/DpNotifyContainer'
 import initGlobalEventListener from '@DpJs/lib/GlobalEventListener'
-import { initStore } from './store/initStore'
+import { initStore } from '@DpJs/store/core/initStore'
 import initToggleSideMenu from '@DpJs/lib/ToggleSideMenu'
 import initUserbox from '@DpJs/lib/touchFriendlyUserbox'
 import { loadLibs } from '@DpJs/lib/loadLibs'
